+++ src/main/java/scratch/peter/nshmp/NSHMP_UCERF3_ERF.java
@@ -9,10 +9,10 @@ import org.opensha.commons.geo.Region;
 import org.opensha.sha.earthquake.ProbEqkSource;
 import org.opensha.sha.earthquake.faultSysSolution.FaultSystemRupSet;
 import org.opensha.sha.earthquake.faultSysSolution.FaultSystemSolution;
+import org.opensha.sha.earthquake.faultSysSolution.modules.GridSourceProvider;
 import org.opensha.sha.earthquake.param.BackgroundRupType;
 
 import scratch.UCERF3.erf.FaultSystemSolutionERF;
-import scratch.UCERF3.griddedSeismicity.GridSourceProvider;
 
 import com.google.common.collect.Lists;
